+++ contacts/src/main/java/com/ingsoftware/contacts/repositories/ContactRepository.java
@@ -8,14 +8,14 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContactRepository extends JpaRepository<Contact, Integer> {
+public interface ContactRepository extends JpaRepository<Contact, Long> {
 
   @Query("SELECT c FROM Contact c WHERE c.user.id=?1")
-  List<Contact> findAllByUser(int id);
+  List<Contact> findAllByUser(long id);
 
   @Query("SELECT c FROM Contact c WHERE c.user.id=?1 AND c.id=?2")
-  Contact findByIdAndUser(int idUser, int idContact);
+  Contact findByIdAndUser(long idUser, long idContact);
 
   @Query("DELETE FROM Contact c WHERE c.user.id=?1 AND c.id=?2")
-  String deleteByUserAndId(int idUser, int idContact);
+  String deleteByUserAndId(long idUser, long idContact);
 }
